requirements.txt
# Core inference
openai-whisper>=20231117
numpy>=1.24
gradio>=4.0

# Audio
soundfile>=0.12

# Progress store encryption
cryptography>=41.0

# Evaluation notebook
scikit-learn>=1.3
matplotlib>=3.7
pandas>=2.0
jupyter>=1.0
ipykernel>=6.0

# Image (visual grounding)
Pillow>=10.0

# Optional: GGUF language head (install separately if using LLM feedback)
# llama-cpp-python>=0.2.0

# Optional: OWLViT visual grounding (heavier, skip for minimal footprint)
# transformers>=4.35
# torch>=2.0
